Removed all dependencies from .ui files (.ui files are not deleted, as it`s easier to recompile them back to .py after changes made in Qt Designer)
