+++ .gitignore
@@ -1,4 +1,4 @@
 venv
 .vscode
-__pycache__
 .idea
+__pycache__
