+++ ui/Widgets/CreateCurveWidget.ui
