+++ ui/Widgets/CreateRotationFigureWidget.ui
